Extract tracks manager helper functions
